requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
python-dotenv==1.0.0
